+++ nn_dataflow/Scheduling.py
@@ -18,11 +18,11 @@ You should have received a copy of the Modified BSD-3 License along with this
 program. If not, see <https://opensource.org/licenses/BSD-3-Clause>.
 """
 
-import math
 from collections import OrderedDict, namedtuple
 
 from . import LoopBlocking
 from . import Partition
+from . import Util
 from .Cost import Cost
 from .DataLayout import DataLayout
 from .Layer import Layer
@@ -51,8 +51,7 @@ class SchedulingCondition(namedtuple('SchedulingCondition',
 
 
 class SchedulingResult(namedtuple('SchedulingResult',
-                                  ['total_cost',
-                                   'dict_loop',
+                                  ['dict_loop',
                                    'dict_part',
                                    'ofmap_layout',
                                   ])):
@@ -68,10 +67,37 @@ class SchedulingResult(namedtuple('SchedulingResult',
             raise TypeError('SchedulingResult: dict_loop and dict_part '
                             'must be OrderedDict instances.')
         if not isinstance(ntp.ofmap_layout, DataLayout):
-            raise TypeError('SchedulingCondition: ofmap_layout must be '
+            raise TypeError('SchedulingResult: ofmap_layout must be '
                             'a DataLayout instance.')
 
         return ntp
+
+    @property
+    def total_cost(self):
+        ''' Get the total cost. '''
+        return self.dict_loop['cost'] + self.dict_part['cost']
+
+    @property
+    def total_time(self):
+        ''' Get the dataflow total time. '''
+        return self.dict_loop['time']
+
+    @property
+    def total_ops(self):
+        ''' Get the total ops. '''
+        # dict_loop stats are over all nodes.
+        return self.dict_loop['ops']
+
+    @property
+    def total_accesses(self):
+        ''' Get the total accesses at all memory hierarchies as a list. '''
+        # dict_loop stats are over all nodes.
+        return [sum(acc) for acc in self.dict_loop['access']]
+
+    @property
+    def total_noc_hops(self):
+        ''' Get the total NoC hops. '''
+        return sum(self.dict_part['total_nhops'])
 
 
 class Scheduling(object):
@@ -115,6 +141,13 @@ class Scheduling(object):
         if not ifmap_layout.is_in_region(mem_region_src):
             raise ValueError('Scheduling: ifmap layout contains invalid '
                              'source memory nodes.')
+        cifrng = ifmap_layout.frmap.complete_fmap_range()
+        if cifrng.size('b') != self.batch_size \
+                or cifrng.size('n') != self.layer.nifm \
+                or not self.layer.is_valid_padding_sifm([cifrng.size('h'),
+                                                         cifrng.size('w')]):
+            raise ValueError('Scheduling: ifmap layout does not match '
+                             'input layer.')
 
         # Filter nodes. All memory nodes can store filters. Deduplicate.
         filter_node_coord_list = [c for c in mem_region_src.node_iter()] \
@@ -133,16 +166,13 @@ class Scheduling(object):
             unit_nhops = Partition.part_layer_unit_nhops(
                 self.layer, self.batch_size, part, filter_node_coord_list,
                 ifmap_layout, ofmap_layout, options)
-            if math.isinf(sum(unit_nhops)):
-                continue
 
             # Explore single-node schedules.
             for lbs in self.schedule_search_per_node(
                     part, condition.resource, options):
 
                 # Make scheduling result.
-                r = self._get_result(lbs, part, unit_nhops, ofmap_layout,
-                                     condition, options)
+                r = self._get_result(lbs, part, unit_nhops, ofmap_layout)
                 tops.append(r)
 
         # Pick the top n.
@@ -152,18 +182,16 @@ class Scheduling(object):
         # Initial occupation also applies to layer.
         total_layer_ops = self.layer.total_ops(self.batch_size)
         for t in tops:
-            sum_part_layer_ops = t.dict_loop['ops'] \
-                    * condition.resource.dim_nodes.size()
-            assert abs(float(total_layer_ops) / sum_part_layer_ops - 1) < 1e-4
+            actual_layer_ops = t.dict_loop['ops']
+            assert Util.isclose(total_layer_ops, actual_layer_ops, rel_tol=1e-4)
 
         # Check ofmap layout matches the layer.
         for t in tops:
             cofrng = t.ofmap_layout.frmap.complete_fmap_range()
-            b_rng, n_rng, h_rng, w_rng = cofrng.beg_end()
-            assert b_rng[1] - b_rng[0] == self.batch_size \
-                    and n_rng[1] - n_rng[0] == self.layer.nofm \
-                    and h_rng[1] - h_rng[0] == self.layer.hofm \
-                    and w_rng[1] - w_rng[0] == self.layer.wofm
+            assert cofrng.size('b') == self.batch_size \
+                    and cofrng.size('n') == self.layer.nofm \
+                    and cofrng.size('h') == self.layer.hofm \
+                    and cofrng.size('w') == self.layer.wofm
 
         return list(tops)
 
@@ -220,34 +248,25 @@ class Scheduling(object):
 
         return top_lbs_list
 
-    def _get_result(self, lbs, part, unit_nhops, ofmap_layout, condition,
-                    options):
+    def _get_result(self, lbs, part, unit_nhops, ofmap_layout):
         '''
         Make the schedule result from loop blocking and partitioning.
         '''
-        del options  # unused
 
         # Loop blocking.
-        cost_loop = lbs.get_cost(self.cost)
-        dict_loop = OrderedDict([('cost', cost_loop)])
-        dict_loop.update(lbs.get_scheme_dict())
+        dict_loop = lbs.get_scheme_dict(self.cost)
 
         # Partitioning.
-        total_nhops = [unh * f
-                       for unh, f
-                       in zip(unit_nhops,
-                              lbs.get_top_level_fetch())]
+        total_nhops = [unh * f for unh, f
+                       in zip(unit_nhops, lbs.get_top_level_fetch())]
         cost_part = self.cost.noc_hop * sum(total_nhops)
         dict_part = OrderedDict([('cost', cost_part),
+                                 ('num_nodes', part.size()),
                                  ('total_nhops', total_nhops),
                                  ('part', part.__dict__),
                                  ('unit_nhops', unit_nhops)])
 
-        # Result.
-        total_cost = cost_loop * condition.resource.dim_nodes.size() + cost_part
-
-        return SchedulingResult(total_cost=total_cost,
-                                dict_loop=dict_loop,
+        return SchedulingResult(dict_loop=dict_loop,
                                 dict_part=dict_part,
                                 ofmap_layout=ofmap_layout)